requirements.txt
django>=3.0.3
pytest-cov>=2.8.1 
pytest-django>=3.7.0 
tox>=3.14.3 
flake8>=3.7.9 
flake8-docstrings>=1.5.0
psycopg2>=2.8.4
Sphinx==2.3.1
requests>=2.22.0
beautifulsoup4>=4.8.2
model-bakery>=1.0.2
django-crispy-forms>=1.8.1
django-filter>=2.2.0
Fabric3>=1.14.post1
sentry-sdk>=0.14.1